+++ src/components/Item.tsx
@@ -1,22 +1,16 @@
-import { ReactElement, ReactNode } from 'react';
+import { ReactElement } from 'react';
 import { Item, ItemProps } from 'react-stately';
 
-import { Styles } from '../tasty';
+import { CubeItemBaseProps } from './content/ItemBase/ItemBase';
 
-export interface CubeItemProps<T> extends ItemProps<T> {
-  qa?: string;
-  description?: ReactNode;
-  descriptionPlacement?: 'inline' | 'block' | 'auto';
-  icon?: ReactNode | 'checkbox';
-  prefix?: ReactNode;
-  suffix?: ReactNode;
-  rightIcon?: ReactNode;
-  styles?: Styles;
+export interface CubeItemProps<T>
+  extends ItemProps<T>,
+    Omit<CubeItemBaseProps, 'children'> {
   onAction?: () => void;
   wrapper?: (item: ReactElement) => ReactElement;
   [key: string]: any;
 }
 
-const _Item = Item as unknown as <T>(props: CubeItemProps<T>) => ReactElement;
+const _Item = Item as <T>(props: CubeItemProps<T>) => ReactElement;
 
 export { _Item as Item };